+++ roqoqo/src/measurements/classical_register_measurement.rs
@@ -19,6 +19,7 @@ use std::collections::HashMap;
 ///
 /// Runs a sequence of circuits and returns the classical registers written during circuit execution.
 #[derive(Debug, PartialEq, Clone)]
+#[cfg_attr(feature = "json_schema", derive(schemars::JsonSchema))]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
 pub struct ClassicalRegister {
     /// Constant Circuit that is executed before each Circuit in circuits.
@@ -67,7 +68,7 @@ impl Measure for ClassicalRegister {
         }
         let new_constant_circuit = match &self.constant_circuit {
             None => None,
-            Some(c) => Some(c.substitute_parameters(&mut calculator)?),
+            Some(c) => Some(c.substitute_parameters(&calculator)?),
         };
         let mut new_circuits = Vec::new();
         for circ in self.circuits.iter() {
@@ -75,7 +76,7 @@ impl Measure for ClassicalRegister {
             for (name, val) in substituted_parameters.iter() {
                 calculator.set_variable(name, *val)
             }
-            new_circuits.push(circ.substitute_parameters(&mut calculator)?)
+            new_circuits.push(circ.substitute_parameters(&calculator)?)
         }
         Ok(Self {
             constant_circuit: new_constant_circuit,