+++ roqoqo/src/operations/multi_qubit_gate_operations.rs
@@ -40,6 +40,7 @@ use serde::{Deserialize, Serialize};
     roqoqo_derive::Rotate,
 )]
 #[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "json_schema", derive(schemars::JsonSchema))]
 pub struct MultiQubitMS {
     /// The qubits involved in the multi qubit Molmer-Sorensen gate.
     qubits: Vec<usize>,
@@ -107,6 +108,7 @@ impl OperateMultiQubitGate for MultiQubitMS {
     roqoqo_derive::Rotate,
 )]
 #[cfg_attr(feature = "serialize", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "json_schema", derive(schemars::JsonSchema))]
 pub struct MultiQubitZZ {
     /// The qubits involved in the multi qubit Molmer-Sorensen gate.
     qubits: Vec<usize>,