+++ roqoqo/tests/integration/measurements/classical_register_measurement.rs
@@ -14,10 +14,13 @@
 
 use std::collections::HashMap;
 
+use jsonschema::{Draft, JSONSchema};
 use roqoqo::measurements::ClassicalRegister;
 use roqoqo::operations;
 use roqoqo::prelude::*;
 use roqoqo::Circuit;
+use schemars::schema_for;
+
 #[test]
 fn test_returning_circuits() {
     let mut circs: Vec<Circuit> = vec![Circuit::new()];
@@ -114,4 +117,40 @@ fn test_substitute_parameters_fail() {
     map.insert("teta2".to_string(), 1.0);
     let br_substitutes = br.substitute_parameters(map);
     assert!(br_substitutes.is_err());
+}
+
+#[cfg(feature = "json_schema")]
+#[test]
+fn test_registers_json() {
+    // setting up classical registers measurement
+    let mut circs: Vec<Circuit> = Vec::new();
+    let mut circ1 = Circuit::new();
+    let mut circ1_subs = Circuit::new();
+    circ1 += operations::RotateX::new(0, "theta".into());
+    circ1_subs += operations::RotateX::new(0, 0.0.into());
+    let mut circ2 = Circuit::new();
+    let mut circ2_subs = Circuit::new();
+    circ2 += operations::RotateZ::new(0, "theta2".into());
+    circ2_subs += operations::RotateZ::new(0, 1.0.into());
+    circs.push(circ1);
+    let br = ClassicalRegister {
+        constant_circuit: Some(circ2),
+        circuits: circs.clone(),
+    };
+
+    // Serialize Measurement
+    let test_json = serde_json::to_string(&br).unwrap();
+    let test_value: serde_json::Value = serde_json::from_str(&test_json).unwrap();
+
+    // Create JSONSchema
+    let test_schema = schema_for!(ClassicalRegister);
+    let schema = serde_json::to_string(&test_schema).unwrap();
+    let schema_value: serde_json::Value = serde_json::from_str(&schema).unwrap();
+    let compiled_schema = JSONSchema::options()
+        .with_draft(Draft::Draft7)
+        .compile(&schema_value)
+        .unwrap();
+
+    let validation_result = compiled_schema.validate(&test_value);
+    assert!(validation_result.is_ok());
 }