+++ CHANGELOG.md
@@ -8,6 +8,27 @@ This changelog track changes to the qoqo project starting at version 0.5.0
 
 ### Added
 
+* Semver-style version checking for Circuit serialization. In beta mode (0.y.z) minor version must match (y_library == y_data) in release mode (x.y.z) major version must match (x_library == x_data) and minor version of library must exceed minor version of data (y_library >= y_data).
+* `json_schema` implementing `JsonSchema` from schemars for roqoqo data structures.
+* * Unit tests to validate `json_schema` added for Circuit, QuantumProgram and measurements.
+* `roqoqo-test` extended by two new functions for stochastic gate tests: `construct_random_circuit` and `add_random_multi_qubit_gate`. 
+* A conversion function `to_single_qubit_gate` has been implemented for the OperateSingleQubitGate trait.
+* The multiplication function `mul` added to the python interface, i.e. qoqo, for Single Qubit Gates.
+
+### Changed
+
+* The multiplication function `mul` for single qubit gates has been updated so that the result is always normalized.
+* `qoqo/examples` has been moved to the new github repository `qoqo_examples` which also includes qoqo examples in Rust now.
+* Dependencies have been updated to `qoqo_calculator = 0.7` and `pyo3 = 0.16`. Qoqo python interface has been migrated from #[pyproto] to #[pymethods]. Mutable qoqo_calculator:Calculator has been changed to unmutable where possible after the upgrade to qoqo_calculator version 0.7.
+
+### Changed
+* BasisRotation and CheatedBasisRotation measurements renamed to PauliZProduct and CheatedPauliZProduct measurement to reflect that this is the measurement of the PauliProduct in the z-basis.
+
+
+### Changed
+* BasisRotation and CheatedBasisRotation measurements renamed to PauliZProduct and CheatedPauliZProduct measurement to reflect that this is the measurement of the PauliProduct in the z-basis.
+
+
 ## 0.10.0
 
 ### Fixed 0.10.0